fix: change input parameter type from float32 to float64 in update_toilet_data

+++ server/api/update_toilet_data.go
@@ -50,7 +50,7 @@ func UpdateToiletData(c *gin.Context) {
 		return
 	}
 
-	toilet := model.Toilet{Id: fmt.Sprint(toiletId), Percentage: float32(percentage), Location: fmt.Sprint(location), State: model.ToiletState(fmt.Sprint(state))}
+	toilet := model.Toilet{Id: fmt.Sprint(toiletId), Percentage: float64(percentage), Location: fmt.Sprint(location), State: model.ToiletState(fmt.Sprint(state))}
 	isToiletExist := model.IsToiletExists(toilet.GetId())
 	if !isToiletExist {
 		c.JSON(http.StatusBadRequest, gin.H{
